Add files via upload
This project involves creating a functional clone of the Amazon web page, focusing on key elements of e-commerce design and functionality. The goal is to develop an interactive and visually appealing website that simulates the user experience of browsing and shopping on Amazon. This project will cover front-end development, back-end integration, and database management to deliver an end-to-end solution for an online marketplace.